code-standardizer-engine/requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
supabase
PyPDF2
jina
pypdf